Collapse empty elements when sorting poms
The release plugin collapses elements, while the sortpom plugin
expanded them. So there was lots of back and forth between
plugins. Since the release plugin does not allow to customize the
behaviour, we switch the sortpom plugin to stop expanding empty items.
This will reduce the back and forth between plugins considerably.