Bring release commit messages closer to our standard format
The `[maven-release-plugin] ` prefix does not add much value to the
commit message, so we drop it. Since Maven treats empty strings as
null, we cannot get rid of the prefix completely. But by setting it to
`Release: `, we're closer to the recommended format. So we switch to
that.

Unfortunately, Maven per default trims trailing whitespace, which
would kill the space after the colon and give us even uglier commit
messages. By requiring Maven 3.1.0, we can use xml's white-space
preserving. But as not everyone is on at least 3.1.0 (e.g.: Some of
our Jenkins jobs use older Maven's) and we only need Maven 3.1.0
during releasing, we require it only for releasing. That however means
we need to run executions by id, which only arrived with Maven 3.3.1.
So we have to require that version.
However, with this setup, there is no Maven requirement for building.
Users can build (package, verify) with older Mavens without issues.
But for releasing, Maven 3.3.1 is required.

Change-Id: I204c7c50ee62950aa72442f52097eacb630e4170